requirements.txt
Django~=4.1.1
stripe~=4.1.0
python-dotenv~=0.21.0